+++ modules/AWSPowerShell/Cmdlets/AWSSupport/Basic/Get-ASAService-Cmdlet.cs
@@ -44,7 +44,7 @@ namespace Amazon.PowerShell.Cmdlets.ASA
     /// You must have a Business, Enterprise On-Ramp, or Enterprise Support plan to use the
     /// Amazon Web Services Support API. 
     /// </para></li><li><para>
-    /// If you call the Amazon Web Services Support API from an account that does not have
+    /// If you call the Amazon Web Services Support API from an account that doesn't have
     /// a Business, Enterprise On-Ramp, or Enterprise Support plan, the <code>SubscriptionRequiredException</code>
     /// error message appears. For information about changing your support plan, see <a href="http://aws.amazon.com/premiumsupport/">Amazon
     /// Web Services Support</a>.
@@ -63,9 +63,10 @@ namespace Amazon.PowerShell.Cmdlets.ASA
         #region Parameter Language
         /// <summary>
         /// <para>
-        /// <para>The ISO 639-1 code for the language in which Amazon Web Services provides support.
-        /// Amazon Web Services Support currently supports English ("en") and Japanese ("ja").
-        /// Language parameters must be passed explicitly for operations that take them.</para>
+        /// <para>The language in which Amazon Web Services Support handles the case. Amazon Web Services
+        /// Support currently supports English ("en") and Japanese ("ja"). You must specify the
+        /// ISO 639-1 code for the <code>language</code> parameter if you want support in that
+        /// language.</para>
         /// </para>
         /// </summary>
         [System.Management.Automation.Parameter(Position = 1, ValueFromPipelineByPropertyName = true)]