+++ modules/AWSPowerShell/Cmdlets/AWSSupport/Basic/Request-ASATrustedAdvisorCheckRefresh-Cmdlet.cs
@@ -31,20 +31,28 @@ namespace Amazon.PowerShell.Cmdlets.ASA
     /// Refreshes the Trusted Advisor check that you specify using the check ID. You can get
     /// the check IDs by calling the <a>DescribeTrustedAdvisorChecks</a> operation.
     /// 
-    ///  <note><para>
+    ///  
+    /// <para>
     /// Some checks are refreshed automatically. If you call the <code>RefreshTrustedAdvisorCheck</code>
     /// operation to refresh them, you might see the <code>InvalidParameterValue</code> error.
-    /// </para></note><para>
+    /// </para><para>
     /// The response contains a <a>TrustedAdvisorCheckRefreshStatus</a> object.
     /// </para><note><ul><li><para>
     /// You must have a Business, Enterprise On-Ramp, or Enterprise Support plan to use the
     /// Amazon Web Services Support API. 
     /// </para></li><li><para>
-    /// If you call the Amazon Web Services Support API from an account that does not have
+    /// If you call the Amazon Web Services Support API from an account that doesn't have
     /// a Business, Enterprise On-Ramp, or Enterprise Support plan, the <code>SubscriptionRequiredException</code>
     /// error message appears. For information about changing your support plan, see <a href="http://aws.amazon.com/premiumsupport/">Amazon
     /// Web Services Support</a>.
-    /// </para></li></ul></note>
+    /// </para></li></ul></note><para>
+    /// To call the Trusted Advisor operations in the Amazon Web Services Support API, you
+    /// must use the US East (N. Virginia) endpoint. Currently, the US West (Oregon) and Europe
+    /// (Ireland) endpoints don't support the Trusted Advisor operations. For more information,
+    /// see <a href="https://docs.aws.amazon.com/awssupport/latest/user/about-support-api.html#endpoint">About
+    /// the Amazon Web Services Support API</a> in the <i>Amazon Web Services Support User
+    /// Guide</i>.
+    /// </para>
     /// </summary>
     [Cmdlet("Request", "ASATrustedAdvisorCheckRefresh", SupportsShouldProcess = true, ConfirmImpact = ConfirmImpact.Medium)]
     [OutputType("Amazon.AWSSupport.Model.TrustedAdvisorCheckRefreshStatus")]