+++ modules/AWSPowerShell/Cmdlets/LicenseManagerLinuxSubscriptions/AWS.Tools.LicenseManagerLinuxSubscriptions.Aliases.psm1
@@ -1,0 +1,3 @@
+﻿# LicenseManagerLinuxSubscriptions
+
+Export-ModuleMember -Alias *